update sql definitions

+++ database.sql
@@ -2,7 +2,7 @@
 CREATE TABLE `users` (
 	`id` INT NOT NULL AUTO_INCREMENT,
 	`name` TEXT NOT NULL,
-	`balance` DECIMAL NOT NULL,
+	`balance` FLOAT NOT NULL,
 	`cards` TEXT,
 	PRIMARY KEY (`id`)
 );
